+++ release/fv/fv_dene_dzage/source/help/fv_dene_dzage.php
@@ -1,6 +1,6 @@
 <?php 
-  $pagename = 'Dene Dzage Keyboard Help';
-  $pagetitle = 'Dene Dzage Keyboard Help';
+  $pagename = 'Danezāgéʼ Keyboard Help';
+  $pagetitle = 'Danezāgéʼ Keyboard Help';
   $pagestyle = <<<END
   
 #wikitext table { margin: 12px; }
@@ -13,91 +13,153 @@ END;
 ?>
 
 <p style='margin: 16px 0 0 0'>
-This keyboard is designed for the <b>Dene Dzage</b> language of the Western Subarctic region of Canada.
+    This keyboard is designed for the <b>Danezāgéʼ</b> language of the Western Subarctic region of Canada.
 </p>
-<p>If square boxes are displayed instead of characters when using this keyboard (and in the keyboard layouts below), please read our <a href="/troubleshooting/#boxes">troubleshooting guide</a>.
+<p>If square boxes are displayed instead of characters when using this keyboard (and in the keyboard layouts below),
+    please read our <a href="/troubleshooting/#boxes">troubleshooting guide</a>.
 </p>
 
 <h2>Desktop layout</h2>
-<p>The Dene Dzage keyboard uses the English keyboard as a base, with added key combinations in order to create special characters. See <a href="file:fv_dene_dzage_combinations.pdf">this document</a> for all the special key combinations.</p>
+<p>The Danezāgéʼ keyboard uses the English keyboard as a base, with added key combinations in order to create special
+    characters. See <a href="file:fv_dene_dzage_combinations.pdf">this document</a> for all the special key
+    combinations.</p>
 </p>
-<p>Typing a single punctuation mark followed by a vowel (without a space in between) will type the vowel with a high, rising, or falling tone, the nasal hook, or a long vowel.</p>
-  <ul>
-    <li><b>Backquote `</b> plus a vowel becomes a high toned vowel</li>
-      <ul>
-        <li><b>`a</b> becomes <b>á</b></li>
-      </ul>
-    <li><b>Semicolon ;</b> plus a vowel becomes a long vowel</li>
-      <ul>
-        <li><b>;e</b> becomes <b>ē</b></li>
-      </ul>
-    <li><b>Left Bracket [</b> plus a vowel becomes a vowel with rising tone</li>
-      <ul>
-        <li><b>[i</b> becomes <b>ǐ</b></li>
-      </ul>
-    <li><b>Right Bracket ]</b> plus a vowel becomes a vowel with falling tone</li>
-      <ul>
-        <li><b>[o</b> becomes <b>ô</b></li>
-      </ul>
-    <li><b>Comma ,</b> plus a vowel becomes a vowel with nasal hook</li>
-      <ul>
-        <li><b>,u</b> becomes <b>ų</b></li>
-      </ul>
-    <li><b>Equal Sign =</b> plus a vowel becomes the combined long vowel with high tone</li>
-      <ul>
-        <li><b>=a</b> becomes <b>ā́</b></li>
-      </ul>
-  </ul>
-<p>Typing two of a punctuation mark followed by a vowel (without a space in between) will type a a high, rising, or falling tone, or a long vowel, <b>with</b> a nasal hook.</p>
-  <ul>
-    <li><b>Two Backquotes ``</b> plus a vowel becomes a nasal high toned vowel</li>
-      <ul>
-        <li><b>``e</b> becomes <b>ę́</b></li>
-      </ul>
-    <li><b>Two Semicolons ;;</b> plus a vowel becomes a nasal long vowel</li>
-      <ul>
-        <li><b>;;o</b> becomes <b>ǭ</b></li>
-      </ul>
-    <li><b>Two Left Brackets [[</b> plus a vowel becomes a nasal rising tone vowel</li>
-      <ul>
-        <li><b>[[a</b> becomes <b>ą̌ </b></li>
-      </ul>
-    <li><b>Two Right Brackets ]]</b> plus a vowel becomes a nasal falling tone vowel</li>
-      <ul>
-        <li><b>]]e</b> becomes <b>ę̂</b></li>
-      </ul>
-    <li><b>Two Equal Signs ==</b> plus a vowel becomes a long, nasal high tone vowel</li>
-      <ul>
-        <li><b>==o</b> becomes <b>ǭ́</b>
-      </ul>
-  </ul>
+<p>Typing a single punctuation mark followed by a vowel (without a space in between) will type the vowel with a high,
+    rising, or falling tone, the nasal hook, or a long vowel.</p>
+<ul>
+    <li><b>Backquote `</b> plus a vowel becomes a high toned vowel
+        <ul>
+            <li><b>`a</b> becomes <b>á</b></li>
+        </ul>
+    </li>
+    <li><b>Semicolon ;</b> plus a vowel becomes a long vowel
+        <ul>
+            <li><b>;e</b> becomes <b>ē</b></li>
+        </ul>
+    </li>
+    <li><b>Left Bracket [</b> plus a vowel becomes a vowel with rising tone
+        <ul>
+            <li><b>[i</b> becomes <b>ǐ</b></li>
+        </ul>
+    </li>
+    <li><b>Right Bracket ]</b> plus a vowel becomes a vowel with falling tone
+        <ul>
+            <li><b>[o</b> becomes <b>ô</b></li>
+        </ul>
+    </li>
+    <li><b>Comma ,</b> plus a vowel becomes a vowel with nasal hook
+        <ul>
+            <li><b>,u</b> becomes <b>ų</b></li>
+        </ul>
+    </li>
+    <li><b>Equal Sign =</b> plus a vowel becomes the combined long vowel with high tone
+        <ul>
+            <li><b>=a</b> becomes <b>ā́</b></li>
+        </ul>
+    </li>
+</ul>
+<p>Typing two of a punctuation mark followed by a vowel (without a space in between) will type a a high, rising, or
+    falling tone, or a long vowel, <b>with</b> a nasal hook.</p>
+<ul>
+    <li><b>Two Backquotes ``</b> plus a vowel becomes a nasal high toned vowel
+        <ul>
+            <li><b>``e</b> becomes <b>ę́</b></li>
+        </ul>
+    </li>
+    <li><b>Two Semicolons ;;</b> plus a vowel becomes a nasal long vowel
+        <ul>
+            <li><b>;;o</b> becomes <b>ǭ</b></li>
+        </ul>
+    </li>
+    <li><b>Two Left Brackets [[</b> plus a vowel becomes a nasal rising tone vowel
+        <ul>
+            <li><b>[[a</b> becomes <b>ą̌ </b></li>
+        </ul>
+    </li>
+    <li><b>Two Right Brackets ]]</b> plus a vowel becomes a nasal falling tone vowel
+        <ul>
+            <li><b>]]e</b> becomes <b>ę̂</b></li>
+        </ul>
+    </li>
+    <li><b>Two Equal Signs ==</b> plus a vowel becomes a long, nasal high tone vowel
+        <ul>
+            <li><b>==o</b> becomes <b>ǭ́</b>
+        </ul>
+    </li>
+</ul>
 <p>Other Combinations</p>
-  <ul>
+<ul>
     <li><b>ł</b> is typed <b>;l</b></li>
     <li>The Straight Quote <b>'</b> is typed <b>Right-Alt</b> (Windows) or <b>Option</b> (Mac) + <b>Quote Key</b></li>
-  </ul>
+</ul>
 <p>
 </p>
 <h2>Unicode Fonts</h2>
-						<p>The following fonts have been installed on your computer: <strong>Aboriginal Sans, Aboriginal Serif.</strong> Please use these fonts when typing in your language as they have been specially built to represent your language as accurately as possible. </p>
-    <h3><a href="http://www.languagegeek.com/font/fontdownload.html">Download the latest version of Aboriginal Sans and Aboriginal Serif</a></h3>
-    <hr></td>
-  </tr>
+<p>The following fonts have been installed on your computer: <strong>Aboriginal Sans, Aboriginal Serif.</strong> Please
+    use these fonts when typing in your language as they have been specially built to represent your language as
+    accurately as possible. </p>
+<h3><a href="http://www.languagegeek.com/font/fontdownload.html">Download the latest version of Aboriginal Sans and
+        Aboriginal Serif</a></h3>
+</td>
+</tr>
 
 <h2>Notes on touch layout</h2>
 <div id='wikitext'>
-<div class='vspace'></div>
-<table border='1' style='border-collapse:collapse' cellpadding='5' ><tr ><th  align='center'>´</th><td  align='center'>w</td><td  align='center'>e</td><th  align='center'>ˇ</th><td  align='center'>t</td><td  align='center'>y</td><td  align='center'>u</td><td  align='center'>i</td><td  align='center'>o</td><td  align='center'>ł</td></tr>
-<tr ><td  align='center'>a</td><td  align='center'>s</td><td  align='center'>d</td><th  align='center'>˛</th><td  align='center'>g</td><td  align='center'>h</td><td  align='center'>j</td><td  align='center'>k</td><td  align='center'>l</td><td  align='center'>' "</td></tr>
-<tr ><td >&nbsp;</td><td  align='center'>z</td><th  align='center'>¯</th><td  align='center'>c</td><th  align='center'>¯́</th><th  align='center'>^</th><td  align='center'>b</td><td  align='center'>n</td><td  align='center'>m</td><td  colspan='2'>&nbsp;</td></tr>
-</table>
-<div class='vspace'></div><ul><li>keys with a red background are combining accents.
-</li><li>the first character in a cell is the "one-tap" key, any further characters are "hold-select" keys.
-</li></ul>
+    <div class='vspace'></div>
+    <table border='1' style='border-collapse:collapse' cellpadding='5'>
+        <tr>
+            <th align='center'>´</th>
+            <td align='center'>w</td>
+            <td align='center'>e</td>
+            <th align='center'>ˇ</th>
+            <td align='center'>t</td>
+            <td align='center'>y</td>
+            <td align='center'>u</td>
+            <td align='center'>i</td>
+            <td align='center'>o</td>
+            <td align='center'>ł</td>
+        </tr>
+        <tr>
+            <td align='center'>a</td>
+            <td align='center'>s</td>
+            <td align='center'>d</td>
+            <th align='center'>˛</th>
+            <td align='center'>g</td>
+            <td align='center'>h</td>
+            <td align='center'>j</td>
+            <td align='center'>k</td>
+            <td align='center'>l</td>
+            <td align='center'>' "</td>
+        </tr>
+        <tr>
+            <td align='center'>⇧</td>
+            <td align='center'>z</td>
+            <th align='center'>¯</th>
+            <td align='center'>c</td>
+            <th align='center'>¯́</th>
+            <th align='center'>^</th>
+            <td align='center'>b</td>
+            <td align='center'>n</td>
+            <td align='center'>m</td>
+            <td align='center'>⌫</td>
+        </tr>
+        <tr>
+            <td align='center'>123</td>
+            <td align='center'>🌐</td>
+            <td align='center' colspan='5'>space</td>
+            <td align='center'>.</td>
+            <td align='center' colspan='2'>⏎</td>
+        </tr>
+    </table>
+    <div class='vspace'></div>
+    <ul>
+        <li>Keys with a red background are combining accents. On your mobile device, these keys will have green text.
+        </li>
+        <li>The first character in a cell is the "one-tap" key, any further characters are "hold-select" keys.
+        </li>
+        <li>Common punctuation marks can be found by pressing and holding the period key on the right side of the space
+            bar.</li>
+    </ul>
 </div>
 
-<br/>
-
-<h2>Keyboard Layout</h2>
-<div id='osk-tablet' data-states='default shift numeric'>
-</div>
+<br />